requirements.txt
Django==2.1.7
pip==18.1.0
Pillow==5.3.0
djangorestframework==3.9.0
channels==2.1.3
mysqlclient==1.4.1
daphne==2.2.2
honcho==1.0.1
requests==2.20.0
fuzzywuzzy==0.17.0
python-Levenshtein==0.12.0